+++ parser/src/parser.rs
@@ -73,16 +73,19 @@ pub fn parse_statement(source: &str) -> Result<ast::LocatedStatement, String> {
 ///
 /// # Example
 /// ```
+/// extern crate num_bigint;
+/// extern crate rustpython_parser;
+/// use num_bigint::BigInt;
 /// use rustpython_parser::{parser, ast};
 /// let expr = parser::parse_expression("1+2").unwrap();
 ///
 /// assert_eq!(ast::Expression::Binop {
 ///         a: Box::new(ast::Expression::Number {
-///             value: ast::Number::Integer { value: 1 }
+///             value: ast::Number::Integer { value: BigInt::from(1) }
 ///         }),
 ///         op: ast::Operator::Add,
 ///         b: Box::new(ast::Expression::Number {
-///             value: ast::Number::Integer { value: 2 }
+///             value: ast::Number::Integer { value: BigInt::from(2) }
 ///         })
 ///     },
 ///     expr);
@@ -98,6 +101,7 @@ mod tests {
     use super::parse_expression;
     use super::parse_program;
     use super::parse_statement;
+    use num_bigint::BigInt;
 
     #[test]
     fn test_parse_empty() {
@@ -150,7 +154,9 @@ mod tests {
                                     value: String::from("Hello world"),
                                 },
                                 ast::Expression::Number {
-                                    value: ast::Number::Integer { value: 2 },
+                                    value: ast::Number::Integer {
+                                        value: BigInt::from(2)
+                                    },
                                 }
                             ],
                             keywords: vec![],
@@ -181,7 +187,9 @@ mod tests {
                             keywords: vec![ast::Keyword {
                                 name: Some("keyword".to_string()),
                                 value: ast::Expression::Number {
-                                    value: ast::Number::Integer { value: 2 },
+                                    value: ast::Number::Integer {
+                                        value: BigInt::from(2)
+                                    },
                                 }
                             }],
                         },
@@ -201,13 +209,17 @@ mod tests {
                 location: ast::Location::new(1, 1),
                 node: ast::Statement::If {
                     test: ast::Expression::Number {
-                        value: ast::Number::Integer { value: 1 },
+                        value: ast::Number::Integer {
+                            value: BigInt::from(1)
+                        },
                     },
                     body: vec![ast::LocatedStatement {
                         location: ast::Location::new(1, 7),
                         node: ast::Statement::Expression {
                             expression: ast::Expression::Number {
-                                value: ast::Number::Integer { value: 10 },
+                                value: ast::Number::Integer {
+                                    value: BigInt::from(10)
+                                },
                             }
                         },
                     },],
@@ -215,13 +227,17 @@ mod tests {
                         location: ast::Location::new(2, 1),
                         node: ast::Statement::If {
                             test: ast::Expression::Number {
-                                value: ast::Number::Integer { value: 2 },
+                                value: ast::Number::Integer {
+                                    value: BigInt::from(2)
+                                },
                             },
                             body: vec![ast::LocatedStatement {
                                 location: ast::Location::new(2, 9),
                                 node: ast::Statement::Expression {
                                     expression: ast::Expression::Number {
-                                        value: ast::Number::Integer { value: 20 },
+                                        value: ast::Number::Integer {
+                                            value: BigInt::from(20)
+                                        },
                                     },
                                 },
                             },],
@@ -229,7 +245,9 @@ mod tests {
                                 location: ast::Location::new(3, 7),
                                 node: ast::Statement::Expression {
                                     expression: ast::Expression::Number {
-                                        value: ast::Number::Integer { value: 30 },
+                                        value: ast::Number::Integer {
+                                            value: BigInt::from(30)
+                                        },
                                     },
                                 },
                             },]),
@@ -295,10 +313,14 @@ mod tests {
                     value: ast::Expression::Tuple {
                         elements: vec![
                             ast::Expression::Number {
-                                value: ast::Number::Integer { value: 4 }
+                                value: ast::Number::Integer {
+                                    value: BigInt::from(4)
+                                }
                             },
                             ast::Expression::Number {
-                                value: ast::Number::Integer { value: 5 }
+                                value: ast::Number::Integer {
+                                    value: BigInt::from(5)
+                                }
                             }
                         ]
                     }
@@ -441,7 +463,9 @@ mod tests {
                                 }),
                                 op: ast::Comparison::Less,
                                 b: Box::new(ast::Expression::Number {
-                                    value: ast::Number::Integer { value: 5 }
+                                    value: ast::Number::Integer {
+                                        value: BigInt::from(5)
+                                    }
                                 }),
                             },
                             ast::Expression::Compare {
@@ -450,7 +474,9 @@ mod tests {
                                 }),
                                 op: ast::Comparison::Greater,
                                 b: Box::new(ast::Expression::Number {
-                                    value: ast::Number::Integer { value: 10 }
+                                    value: ast::Number::Integer {
+                                        value: BigInt::from(10)
+                                    }
                                 }),
                             },
                         ],
